+++ package.json
@@ -6,7 +6,7 @@
     "type": "module",
     "scripts": {
         "start": "node index.js",
-        "serve":"nodemon index.js"
+        "serve": "nodemon index.js"
     },
     "keywords": [
         "Node",
@@ -17,10 +17,11 @@
     "license": "ISC",
     "dependencies": {
         "chalk": "^5.6.2",
+        "express": "^5.1.0",
         "yargs": "^18.0.0"
     },
     "devDependencies": {
-        "nodemon": "^3.1.10",
-        "@types/yargs": "^17.0.33"
+        "@types/yargs": "^17.0.33",
+        "nodemon": "^3.1.10"
     }
 }
