+++ pages/index.html
@@ -18,9 +18,7 @@
                 <div class="form-group">
                     <input type="text" class="form-control mb-2" name="title" placeholder="Enter your text" required/>
                 </div>
-
                 <button class="btn btn-primary" type="submit">Create</button>
-
             </form>
         </div>
     </body>
